+++ CoreBarcelona/Constants.swift
@@ -97,4 +97,4 @@ public let ERChatMessageUpdatedNotification = NSNotification.Name(rawValue: "ERC
 public let ERChatMessagesDeletedNotification = NSNotification.Name(rawValue: "ERChatMessagesDeletedNotification")
 internal let ERChatRegistryDidLoadNotification = NSNotification.Name(rawValue: "ERChatRegistryDidLoadNotification")
 
-public let ERDefaultMessageQueryLimit: Int64 = 75
+public let ERDefaultMessageQueryLimit: Int = 75
